Extract Streams from documentation as standalone docs
Here's the changes we talked about to pull Streams documentation out to it's own page. Ended up having to do the redirect on URLs that contain anchor tags that match the pattern 'streams_' with JS in the head section of the page because anchor tags aren't available on the server so you can't do rewrite rules based off them.

(also added in a change I thought was already out there to show info about Kafka Summit on the events page)

Author: Derrick Or <derrickor@gmail.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #36 from derrickdoo/streamsStandaloneDocs