KAFKA-6312: Update website documentation for --reset-offsets option, for Kafka consumer groups, introduced in KIP-122
KIP-122 added the ability for kafka-consumer-groups.sh to reset/change consumer offsets, at a fine grained level.

There is documentation on it in the kafka-consumer-groups.sh usage text.

There was no such documentation on the kafka.apache.org website. This change updates the documentation on the website, so that users can read about the functionality without having the tools installed.