2.6.0 release (#277)
Add 2.6.0 release to website. Simplified slightly the upgrade language in the upgrade sections.

This also restores the 2.5.0 upgrade notes lost during previous release, which added these 2.5.0 upgrade notes to the kafka-site’s `25/upgrade.html` file but not the `documentation/upgrade.html` in the kafka repo that generates the kafka-site content.