MINOR: document how we deal with advisories for dependencies
We get questions about advisories in dependencies on a regular basis,
often in the form of security scanner tool output. Because of the high
likelihood of false positives, the ASF policy is not to accept such
reports as security issues without additional analysis: we consider
the mere fact that an advisory exists for a dependency already public
knowledge.

This update makes this clearer to reporters, helps them find out
whether an advisory impacts Kafka for themselvels (by pointing to the
dependency-check suppressions and the issue tracker), and calls on
them to help out with analysis.